Reduced further, fixed bugs

+++ 2018/day13/day13.linq
@@ -12,37 +12,30 @@ void Main()
 {
     // sample
 
-    Sim("sample1").ShouldBe((0, 3));
-    Sim("sample2").ShouldBe((7, 3));
-    Sim("sample3", false).ShouldBe((6, 4));
+    Sim("sample1").firstCollision.ShouldBe((0, 3));
+    Sim("sample2").firstCollision.ShouldBe((7, 3));
+    Sim("sample3").lastCart.ShouldBe((6, 4));
 
     // problem
 
-    var firstCollide = Sim("input");
-    firstCollide.ToString().Dump();
-    firstCollide.ShouldBe((83, 121));
-
-    var lastSurviving = Sim("input", false);
-    lastSurviving.ToString().Dump();
-    lastSurviving.ShouldBe((102, 144));
+    var result = Sim("input");
+    Console.WriteLine($"{result.firstCollision.x},{result.firstCollision.y}");
+    result.firstCollision.ShouldBe((83, 121));
+    Console.WriteLine($"{result.lastCart.x},{result.lastCart.y}");
+    result.lastCart.ShouldBe((102, 144));
 }
 
-class Cart : IComparable<Cart>
+class Cart
 {
     public int X, Y;
     public char C;
     public int AI;
 
     public Cart(int x, int y, char dir) { X = x; Y = y; C = dir; }
-
-    public int CompareTo(Cart other)
-    {
-        var compare = Y.CompareTo(other.Y);
-        return compare == 0 ? X.CompareTo(other.X) : compare;
-    }
 }
 
-(int x, int y) Sim(string name, bool findFirstCollision = true)
+((int x, int y) firstCollision, (int x, int y) lastCart)
+Sim(string name)
 {
     var lines = File.ReadAllLines($"{scriptDir}/{name}.txt");
     var maxLine = lines.Max(l => l.Length);
@@ -50,35 +43,32 @@ class Cart : IComparable<Cart>
     var state = new char[maxLine, lines.Length];
     var carts = new List<Cart>();
     
-    var outPath = $"{scriptDir}/{name}.out";
-    if (File.Exists(outPath))
-        File.Delete(outPath);
-    
+    const string indexers = "^>v<";
+
     void Render(int frame)
     {
+        var outPath = $"{scriptDir}/{name}.out";
+        if (frame == 0 && File.Exists(outPath))
+            File.Delete(outPath);
+
+        // TODO: GIF!
         if (lines.Length > 20)
             return;
         
-        var dup = (char[,])state.Clone();
-        foreach (var c in carts)
-        {
-            if ("<^>v".Contains(dup[c.X, c.Y]))
-                dup[c.X, c.Y] = 'X';
-            else
-                dup[c.X, c.Y] = c.C;
-        }
-
         using (var f = File.AppendText(outPath))
         {
             f.WriteLine($"***[{frame}]***");
             f.WriteLine();
-            
+
+            var dup = (char[,])state.Clone();
+            foreach (var c in carts)
+                dup[c.X, c.Y] = c.C;
+
             for (var y = 0; y < lines.Length; ++y)
             {
-                var sb = new StringBuilder();
                 for (var x = 0; x < maxLine; ++x)
-                    sb.Append(dup[x, y] == 0 ? ' ' : dup[x, y]);
-                f.WriteLine(sb.ToString());
+                    f.Write(dup[x, y] == 0 ? ' ' : dup[x, y]);
+                f.WriteLine();
             }
             f.WriteLine();
 
@@ -89,17 +79,15 @@ class Cart : IComparable<Cart>
         }
     }
 
-    var cartToState = (from:"^v<>", to:"|-");
-
     for (var y = 0; y < lines.Length; ++y)
     {
         var line = lines[y];
         for (var x = 0; x < line.Length; ++x)
         {
-            var found = cartToState.from.IndexOf(line[x]);
+            var found = indexers.IndexOf(line[x]);
             if (found >= 0)
             {
-                state[x, y] = cartToState.to[found / 2];
+                state[x, y] = "|-|-"[found];
                 carts.Add(new Cart(x, y, line[x]));
             }
             else
@@ -107,35 +95,31 @@ class Cart : IComparable<Cart>
         }
     }
 
-    var indexers = "^>v<";
     var moveSpec = (dx: new[] { 0, 1, 0, -1 }, dy: new[] { -1, 0, 1, 0 });
     var turnSpec = (spec: new[] { '/', '\\', 0, 2 }, next: new[] { ">^<v", "<v>^", "<^>v", ">v<^" });
 
-    for (var frame = 0;;++frame)
+    var collisions = new List<(int x, int y)>();
+
+    for (var frame = 0; ; ++frame)
     {
-        carts.Sort();
+        carts = carts.OrderBy(c => c.Y).ThenBy(c => c.X).ToList();
+
         Render(frame);
         
         for (var i = 0; i < carts.Count; ++i)
             if (carts[i].C == 'X')
                 carts.RemoveAt(i--);
+                
+        if (carts.Count < 2)
+            break;
 
-        if (carts.Count == 1 && !findFirstCollision)
-            return (carts[0].X, carts[0].Y);
-
-        for (var i = 0; i < carts.Count; ++i)
+        foreach (var cart in carts.Where(c => c.C != 'X'))
         {
-            var cart = carts[i];
-            if (cart.C == 'X')
-                continue;
-
             var idir = indexers.IndexOf(cart.C);
 
-            // move
             cart.X += moveSpec.dx[idir];
             cart.Y += moveSpec.dy[idir];
 
-            // turn
             var cell = state[cart.X, cart.Y];
             if (cell == '+')
             {
@@ -150,9 +134,12 @@ class Cart : IComparable<Cart>
             foreach (var dup in carts.Where(c => c != cart && c.X == cart.X && c.Y == cart.Y))
             {
                 cart.C = dup.C = 'X';
-                if (findFirstCollision)
-                    return (cart.X, cart.Y);
+                collisions.Add((cart.X, cart.Y));
             }
         }
     }
+    
+    return (
+        collisions.Any() ? collisions[0] : (-1, -1),
+        carts.Any() ? (carts[0].X, carts[0].Y) : (-1, -1));
 }
